display-panel-run-phosh: Allow to take screenshot
Allow to take a screenshot via `-S name`. We use
a delayed callback to give phosh enough time to start
so we don't accidentally bind to another process exposing
the same API. Should that not prove to be robust enough
we can spawn a separate bus.

Closes: https://gitlab.gnome.org/World/Phosh/gmobile/-/issues/10
Part-of: <https://gitlab.gnome.org/World/Phosh/gmobile/-/merge_requests/90>